fix part field input box making admin slow
The field part from stock model was a editable field in the admin, the problem is that the options for that field are all the model parts, so
if you have a lot of them it tries to retrieve them all and tries also that the browser creates a dropwdown with them, which is super expensive.
Now the field is readonly and everything goes fast.